🚧 Needs to implement livechat and notification

+++ src/models/usersModels.ts
@@ -1,39 +1,53 @@
 import { addTimestamps, db } from "./utilsModel";
 import { jumlahKegiatan, kompetensis, users, usersToKompetensis } from "../db/schema";
-import { and, eq, getTableColumns, isNull, sql } from "drizzle-orm";
+import { and, desc, eq, getTableColumns, sql } from "drizzle-orm";
+import { kompetensisColumns } from "./kompetensiModels";
 
 export type UserDataType = typeof users.$inferInsert
 export type UsersToKompetensiDataType = typeof usersToKompetensis.$inferInsert
-
-const { password, ...rest } = getTableColumns(users)
+export const { password, ...userTableColumns } = getTableColumns(users)
 
 export async function fetchAllUser() {
-    return await db.select({ ...rest }).from(users);
+    return await db.select({ ...userTableColumns }).from(users).orderBy(desc(users.nama));
 }
 
 export async function fetchUserByRole(role: 'admin' | 'manajemen' | 'dosen') {
-    return await db.select({ ...rest }).from(users).where(eq(users.role, role))
+    return await db.select({ ...userTableColumns }).from(users).where(eq(users.role, role))
 }
 
 export async function fetchUserComplete(uidUser: string) {
-    let dataUser, dataKompetensi, dataJumlahKegiatan
-    [dataUser] = await db.select({ ...rest }).from(users).where(eq(users.userId, uidUser));
+    let dataUser, dataKompetensi
+    [dataUser] = await db.select({ ...userTableColumns }).from(users).where(eq(users.userId, uidUser));
 
-    const columnKompetensi = getTableColumns(kompetensis)
-    dataKompetensi = await db.select(columnKompetensi).from(usersToKompetensis)
+    if (!dataUser) return
+
+    dataKompetensi = await db.select(kompetensisColumns).from(usersToKompetensis)
         .rightJoin(kompetensis, eq(usersToKompetensis.kompetensiId, kompetensis.kompetensiId))
         .where(eq(usersToKompetensis.userId, uidUser))
 
-    dataJumlahKegiatan = await db.select().from(jumlahKegiatan).where(eq(jumlahKegiatan.userId, uidUser))
     return {
-        user: dataUser,
-        kompetensi: dataKompetensi,
-        jumlah_kegiatan: dataJumlahKegiatan
+        ...dataUser,
+        kompetensi: dataKompetensi
     }
 }
 
-export async function fetchUserOnly(uidUser: string) {
-    const [res] = await db.select({ ...rest }).from(users).where(eq(users.userId, uidUser))
+export async function fetchUserStatistic(uidUser: string, year: number) {
+    const [dataUser] = await db.select({ ...userTableColumns }).from(users).where(eq(users.userId, uidUser));
+    const dataJumlahKegiatan = await db.select().from(jumlahKegiatan).where(year ? and(eq(jumlahKegiatan.userId, uidUser), eq(jumlahKegiatan.year, year)) : eq(jumlahKegiatan.userId, uidUser))
+
+    let totalDalamSetahun: number = 0
+    for (const dat of dataJumlahKegiatan) {
+        totalDalamSetahun += dat.jumlahKegiatan
+    }
+
+    return {
+        totalDalamSetahun,
+        jumlahKegiatan: dataJumlahKegiatan
+    }
+}
+
+export async function fetchUserByUid(uidUser: string) {
+    const [res] = await db.select({ ...userTableColumns }).from(users).where(eq(users.userId, uidUser))
     return res
 }
 
@@ -56,6 +70,7 @@ export async function createUser(data: UserDataType) {
     }
 }
 
+// TODO DO it noww bitch
 export async function addUserKompetensi(uidUser: string, uidKompetensi: string[]) {
 
     let data: UsersToKompetensiDataType[]
@@ -72,10 +87,15 @@ export async function addUserKompetensi(uidUser: string, uidKompetensi: string[]
         }
     })
 
-
-    return await db.select().from(usersToKompetensis).where(eq(usersToKompetensis.userId, uidUser))
+    const userKompetensi = await db.select().from(usersToKompetensis).where(eq(usersToKompetensis.userId, uidUser))
+    const user = await fetchUserByUid(uidUser)
+    return {
+        ...user,
+        kompetensi: userKompetensi
+    }
 }
 
+// TODO DO it noww bitch
 export async function addJumlahKegiatan(uidUser: string, wasDecrement: boolean = false, year: number, month: number) {
     await db.insert(jumlahKegiatan)
         .values(addTimestamps({
@@ -89,18 +109,21 @@ export async function addJumlahKegiatan(uidUser: string, wasDecrement: boolean =
             }, true)
         })
 
-    return await db.select().from(users)
-        .innerJoin(jumlahKegiatan, eq(jumlahKegiatan.userId, users.userId))
-        .where(eq(users.userId, uidUser))
+    const jmlhKgh = await db.select().from(jumlahKegiatan).where(eq(jumlahKegiatan.userId, uidUser))
+    const user = await fetchUserByUid(uidUser)
+    return {
+        ...user,
+        jumlahKegiatan: jmlhKgh
+    }
 }
 
 export async function updateUser(uidUser: string, data: Partial<UserDataType>) {
-    let res;
     await db.update(users).set(addTimestamps(data, true)).where(eq(users.userId, uidUser))
-    res = await db.select().from(users).where(eq(users.userId, uidUser))
+    const [res] = await db.select({ ...userTableColumns }).from(users).where(eq(users.userId, uidUser))
     return res
 }
 
+// TODO DO it noww bitch
 export async function updateKompetensiUser(uidUser: string, uidKompetensi: string[]) {
 
     let data: UsersToKompetensiDataType[]
@@ -116,19 +139,22 @@ export async function updateKompetensiUser(uidUser: string, uidKompetensi: strin
         }, true)
     })
 
-    return await db.select().from(usersToKompetensis).where(eq(usersToKompetensis.userId, uidUser))
+    const kompe = db.select(kompetensisColumns).from(usersToKompetensis).innerJoin(kompetensis, eq(kompetensis.kompetensiId, usersToKompetensis.kompetensiId)).where(eq(usersToKompetensis.userId, uidUser))
+    const user = await fetchUserByUid(uidUser)
+    return {
+        ...user,
+        kompetensi: kompe
+    }
 }
 
 export async function deleteUser(uidUser: string) {
-    const temp = await db.select().from(users).where(eq(users.userId, uidUser))
-    if (temp) {
-        await db.delete(users)
-            .where(eq(users.userId, uidUser))
-    }
+    const user = await fetchUserByUid(uidUser)
+    await db.delete(users).where(eq(users.userId, uidUser))
 
-    return temp
+    return user
 }
 
+// TODO DO it noww bitch
 export async function deleteUserKompetensi(uidUser: string, uidKompetensi: string[]) {
     await db.transaction(async (tx) => {
         for (const uidK of uidKompetensi) {
@@ -137,12 +163,23 @@ export async function deleteUserKompetensi(uidUser: string, uidKompetensi: strin
         }
     })
 
-    return await db.select().from(users).innerJoin(usersToKompetensis, eq(usersToKompetensis.userId, users.userId)).where(eq(users.userId, uidUser))
+    const userKompetensi = await db.select().from(usersToKompetensis).where(eq(usersToKompetensis.userId, uidUser))
+    const user = await fetchUserByUid(uidUser)
+    return {
+        ...user,
+        kompetensi: userKompetensi
+    }
 }
 
+// TODO DO it noww bitch
 export async function deleteJumlahKegiatan(uidUser: string, year: number, month: number) {
     await db.delete(jumlahKegiatan)
         .where(and(eq(jumlahKegiatan.userId, uidUser), eq(jumlahKegiatan.year, year), eq(jumlahKegiatan.month, month)))
 
-    return await db.select().from(users).innerJoin(jumlahKegiatan, eq(jumlahKegiatan.userId, users.userId)).where(eq(jumlahKegiatan.userId, uidUser))
+    const jmlhKgh = await db.select().from(jumlahKegiatan).where(eq(jumlahKegiatan.userId, uidUser))
+    const user = await fetchUserByUid(uidUser)
+    return {
+        ...user,
+        jumlahKegiatan: jmlhKgh
+    }
 }
